refactor: move more scripts back to scripts/ directory
- Move temporal-dev-server.ts back to scripts/ for easier access
- Move test-yogo-email.ts back to scripts/ for easier access
- Update package.json temporal commands to use scripts/ location
- Update import paths in setup/teardown scripts
- Keep only setup/teardown scripts in __tests__/scripts/temporal/
