+++ package.json
@@ -15,9 +15,9 @@
     "test:connection": "jest __tests__/services/",
     "test:unit": "jest __tests__/unit/ --runInBand",
     "test:integration": "TEST_INTEGRATION=true jest __tests__/integration/ --runInBand",
-    "temporal:start": "tsx __tests__/scripts/temporal/temporal-dev-server.ts start",
-    "temporal:stop": "tsx __tests__/scripts/temporal/temporal-dev-server.ts stop",
-    "temporal:check": "tsx __tests__/scripts/temporal/temporal-dev-server.ts check",
+    "temporal:start": "tsx scripts/temporal-dev-server.ts start",
+    "temporal:stop": "tsx scripts/temporal-dev-server.ts stop",
+    "temporal:check": "tsx scripts/temporal-dev-server.ts check",
     "test:watch": "jest --watch",
     "test:coverage": "jest --coverage",
     "test:debug": "node --inspect-brk node_modules/.bin/jest",
